# apps/server/src/modules/catalog/controllers/rooms-admin.controller.ts
/**
 * Rooms Admin Controller
 *
 * Handles BTC-only room management endpoints.
 * All endpoints require JWT authentication and BTC role.
 *
 * Endpoints:
 * - GET /admin/rooms — list all rooms
 * - POST /admin/rooms — create a new room
 */

import {
  Controller,
  Get,
  Post,
  Patch,
  Param,
  Body,
  Query,
  UseGuards,
  HttpCode,
  HttpStatus,
} from "@nestjs/common";

import { JwtAuthGuard } from "@/modules/iam/guards/jwt-auth.guard";
import { RolesGuard } from "@/modules/iam/guards/roles.guard";
import { RateLimit } from "@/shared/decorators/rate-limit.decorator";
import { Roles } from "@/shared/decorators/roles.decorator";

import { CreateRoomDto } from "../dto/create-room.dto";
import { ListRoomsQueryDto } from "../dto/list-rooms-query.dto";
import { UpdateRoomDto } from "../dto/update-room.dto";
import { RoomsService } from "../services/rooms.service";

@Controller("admin/rooms")
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles("BTC")
@RateLimit([{ tier: "T2", limit: 30, windowMs: 60000 }])
export class RoomsAdminController {
  constructor(private readonly roomsService: RoomsService) {}

  /**
   * Lists all available rooms in the system.
   *
   * Returns all rooms regardless of occupancy status. Room capacity
   * information is included for workshop scheduling validation.
   *
   * Security context: Requires BTC role.
   *
   * @returns Array of room DTOs.
   */
  @Get()
  async listRooms(@Query() query: ListRoomsQueryDto) {
    return this.roomsService.listRooms(query);
  }

  /**
   * Retrieves a single room by its UUID.
   *
   * Security context: Requires BTC role.
   *
   * @param roomId - The UUID of the room to retrieve.
   * @returns The room DTO, or FailResult (ROOM_NOT_FOUND).
   */
  @Get(":roomId")
  async getRoomById(@Param("roomId") roomId: string) {
    return this.roomsService.getRoomById(roomId);
  }

  /**
   * Creates a new room for hosting workshops.
   *
   * Validates input with CreateRoomSchema before persisting.
   *
   * Security context: Requires BTC role.
   *
   * @param body - Room creation payload (name, building?, floor?, capacity, floor_plan_url?, facilities?).
   * @returns The newly created room DTO.
   */
  @Post()
  @HttpCode(HttpStatus.CREATED)
  async createRoom(@Body() dto: CreateRoomDto) {
    return this.roomsService.createRoom(dto);
  }

  /**
   * Updates an existing room's attributes.
   *
   * Only provided fields are updated; omitted fields retain their existing values.
   *
   * Security context: Requires BTC role.
   *
   * @param roomId - The UUID of the room to update.
   * @param body - Partial room update payload.
   * @returns The updated room DTO.
   */
  @Patch(":roomId")
  async updateRoom(
    @Param("roomId") roomId: string,
    @Body() dto: UpdateRoomDto
  ) {
    return this.roomsService.updateRoom(roomId, dto);
  }
}