# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UniHub Workshop API
  version: 1.0.0
  description: "REST API specification for the **UniHub Workshop** system — university\
    \ workshop\nregistration & check-in platform. This document is the authoritative\
    \ contract\nbetween backend (NestJS Modular Monolith) and frontend (Next.js Web\
    \ Portal +\nExpo Mobile App).\n\n## Conventions\n\n- **Versioning:** URL-based,\
    \ prefix `/api/v1`.\n- **Authentication:** JWT Bearer (15-minute access token)\
    \ + refresh token (7 days).\n  Refresh token transport is **client-type specific**\
    \ (see `authentication.md` INV-02):\n  - **Web client:** `refresh_token` is set\
    \ as an `HttpOnly; SameSite=Strict` cookie\n    on `/auth/login` and `/auth/refresh`\
    \ responses. JavaScript cannot read it.\n    Access token is returned in the JSON\
    \ body and kept **in memory only** (not localStorage).\n  - **Mobile client:**\
    \ `refresh_token` is also included in the JSON response body\n    (in addition\
    \ to the cookie) because mobile has no cookie jar — app stores it in\n    Android\
    \ Keystore / iOS Keychain.\n- **Idempotency:** `POST /registrations` and `POST\
    \ /payments` require an\n  `X-Idempotency-Key: <UUID v4>` **HTTP header** (client-generated,\
    \ reuse on retry).\n  Server uses 3-state lifecycle (`in_progress` / `completed`\
    \ / `unresolved`).\n  See ADR-08, ADR-15.\n- **Optimistic Locking:** Admin updates\
...
    \ `VND`).\n"
  contact:
...
  license:
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /rooms/{roomId}:
    get:
      tags:
      - Catalog
      summary: Get room detail (with floor plan)
      operationId: getRoom
      security:
      - {}
      - bearerAuth: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Room detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/rooms:
    get:
      tags:
      - Catalog Admin
      summary: List rooms
      operationId: adminListRooms
      parameters:
      - name: q
        in: query
        description: Search by name (ILIKE)
        schema:
          type: string
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Rooms list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags:
      - Catalog Admin
      summary: Create room
      operationId: adminCreateRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
  /admin/rooms/{roomId}:
    parameters:
    - name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Catalog Admin
      summary: Get room (admin)
      operationId: adminGetRoom
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
    patch:
      tags:
      - Catalog Admin
      summary: Update room
      operationId: adminPatchRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Short-lived access token (15 min). RS256 signed. Decoded claims:

        `sub` (studentId or staff UUID), `role`, `user_type`,

        `allowedWorkshopIds` (checkin_staff only), `iat`, `exp`.

        Verified by public key — no DB lookup at Layer ① (authentication.md INV-01).

        '
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 signature of the request body. Per-gateway secret.
...
  parameters:
...
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor from the previous response's `pagination.next_cursor`
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Resource not found, or hidden by row-level filter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
          description: Always false for error responses
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              description: Stable error code clients can switch on
              example: registration.workshop_full
            message:
              type: string
              description: Human-readable error message safe for client display
              example: Workshop đã đầy chỗ
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
              description: Field-level validation details when applicable
        meta:
          $ref: '#/components/schemas/RequestMeta'
...
    RequestMeta:
      type: object
      description: Metadata attached to every API response for tracing
      required:
      - requestId
      - timestamp
      - apiVersion
      properties:
        requestId:
          type: string
          description: Correlation ID for request tracing
          example: req-abc123
        timestamp:
          type: string
          format: date-time
          description: Response timestamp (RFC 3339)
        apiVersion:
          type: string
          description: API version tag
          example: v1
        processingMs:
          type: integer
          description: Server-side processing duration in milliseconds
          example: 42
    FieldError:
      type: object
      description: Field-level validation error details
      required:
      - field
      - rule
      - message
      properties:
        field:
          type: string
          description: Field path that failed validation
          example: email
        rule:
          type: string
          description: Validation rule identifier
          example: isEmail
        message:
          type: string
          description: Human-readable validation message
          example: Invalid email format
        received:
          description: Raw input value (when safe to expose)
    PaginationMeta:
      type: object
      required:
      - limit
      - hasMore
      properties:
        limit:
          type: integer
        nextCursor:
          type: string
          nullable: true
          description: Pass back as `?cursor=` to fetch the next page
        hasMore:
          type: boolean
        total:
          type: integer
          nullable: true
          description: Only populated for offset-aware admin endpoints
...
    RoomSummary:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        building:
          type: string
          nullable: true
        floor:
          type: integer
          nullable: true
        floorPlanUrl:
          type: string
          nullable: true
    Room:
      allOf:
      - $ref: '#/components/schemas/RoomSummary'
      - type: object
        properties:
          capacity:
            type: integer
          facilities:
            type: object
            additionalProperties: true
            example:
              projector: true
              ac: true
              mic: 2
          createdAt:
            type: string
            format: date-time
    RoomCreateRequest:
      type: object
      required:
      - name
      - capacity
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        building:
          type: string
          maxLength: 100
        floor:
          type: integer
        capacity:
          type: integer
          minimum: 1
        floorPlanUrl:
          type: string
        facilities:
          type: object
          additionalProperties: true
...